add new note and delete functions working

+++ public/js/app.js
@@ -1,7 +1,7 @@
-// const notesAccordian = require("./notesaccordianTL")
 
 $(document).ready(function () {
 
+  // Generates notes to page
   const renderNotes = () => {
 
     $("#note-side-view").empty();
@@ -21,21 +21,22 @@ $(document).ready(function () {
         let noteDate = noteData[i].created_at;
 
         const notesAccordian =
-      `<div class="accordion" id="accordion${noteID}" data-id='${noteID}'>
+      `<div class="accordion" id="accordion${noteID}" data-id="${noteID}">
       <div class="card">
         <div class="card-header" id="heading${noteID}">
           <h2 class="mb-0">
             <button class="btn" type="button" data-toggle="collapse" data-target="#collapse${noteID}" aria-expanded="true" aria-controls="collapse${noteID}">
               ${noteTitle}
             </button>
-            <i class="far fa-trash-alt trash"></i>
+            <i class="far fa-trash-alt trash-icon"></i>
+            <i class="far fa-edit edit-icon"></i>
           </h2>
         </div>
         <div id="collapse${noteID}" class="collapse show" aria-labelledby="heading${noteID}" data-parent="#accordion${noteID}">
           <div class="card-body">
             ${noteBody}
           </div>
-          <div>Created on: ${noteDate}</div>
+          <div class="border-top">Created on: ${noteDate}</div>
         </div>
       </div>
     </div>`
@@ -48,27 +49,72 @@ $(document).ready(function () {
     });
   };
 
-  
+  // const updateNote = () => {
 
-  $(document).on("click", ".trash", function(e) {
+  //   let updateNote = {
+  //     id:
+  //     title: $("#update-title-input").val(),
+  //     note_body: $("#update-body-input").val()
+  //   }
+
+  //   $.ajax({
+  //     url: "/apli/notes",
+  //     method: "POST",
+  //     data: updateNote,
+  //     success: renderNotes()
+  //   })
+  // }
+
+  // Delete notes function
+  const deleteNote = () => {
+    
+    if(confirm("Are you sure you want to delete this note?") === true) {
+      $.ajax({
+        url: "/api/notes",
+        method: "DELETE",
+        data: accordianDiscard
+      }).then(function() {
+        renderNotes();
+        $(`#accordian${accordianDiscard}`).remove();
+      })
+    }
+    else {
+      return false;
+    }
+  }
+
+  // On click event to remove notes
+  $(document).on("click", ".trash-icon", function(e) {
     e.preventDefault();
     console.log("this is working")
-    let accordianDiscard = ($(this)
+    let accordianDiscard = $(this)
       .parent()
       .parent()
       .parent()
-      .attr("data-id")
-    )
-    console.log(accordianDiscard.attr("data-id"));
-  
-    $.ajax({
-      url: "/api/notes",
-      method: "DELETE",
-    }).then(function() {
-      $(`#accordian${accordianDiscard}`).remove();
-    })
+      .parent()
+      .data().id
+    
+    deleteID = {
+      id: accordianDiscard
+    }
+
+    if(confirm("Are you sure you want to delete this note?") === true) {
+      $.ajax({
+        url: "/api/notes",
+        method: "DELETE",
+        data: deleteID
+      }).then(function() {
+        renderNotes();
+        $(`#accordian${accordianDiscard}`).remove();
+      })
+    }
+    else {
+      return false;
+    }
+
   });
 
+  // on click event to save new notes
   $(document).on("click", "#save-new-note", function(e) {
     e.preventDefault();
 
@@ -76,16 +122,13 @@ $(document).ready(function () {
       title: $("#note-title-input").val(),
       note_body: $("#note-body-input").val()
     }
-    // console.log(newNote.title, newNote.note_body);
 
       $.ajax({
         url: "/api/notes",
         method: "POST",
-        data: newNote
-      }).then(newNoteData, function() {
-        console.log(newNoteData.id)
-        console.log(newNoteData.title)
-        console.log(newNoteData.note_body)
+        data: newNote,
+        success: renderNotes()
+      
       });
   });
